Make Dynamic Registration configuration extensible
This change allows the Dynamic Registration configuration to be extended
 with properties that can be set per-installation (As opposed to once
at the top-level for all installs).

This is done by taking the third (currently unused) parameter to the
`lti.DynamicRegistration.register`, and performing a deep merge with the
default registration object, which I assume was the original
intent of that parameter.

```javascript
lti.DynamicRegistration.register(
  req.query.openid_configuration,
  req.query.registration_token, {
    'https://purl.imsglobal.org/spec/lti-tool-configuration': {
      custom_parameters: {
        'custom1': 'value1',
        'custom2': 'value2'
      }
    }
  }
)
```

This fixes issue #144